requirements.txt
flask-sqlalchemy
email_validator
WTForms
Flask-WTF
flask-migrate